requirements.txt
streamlit==1.40.1
fastapi==0.115.5
pydantic==2.9.2
uvicorn==0.32.0
mysql-connector-python==9.1.0
pandas==2.2.3
requests==2.32.3
pytest==8.3.3
